add generic RxAppMock (#433)
* add generic RxAppMock

* fix warnings
